+++ common/src/main/java/com/zimbra/common/account/ZAttrProvisioning.java
@@ -2703,6 +2703,14 @@ public class ZAttrProvisioning {
     public static final String A_carbonioAllowFeedback = "carbonioAllowFeedback";
 
     /**
+     * Whether or not Amavis should skip virus-checking
+     *
+     * @since ZCS 23.5.0
+     */
+    @ZAttr(id=3129)
+    public static final String A_carbonioAmavisDisableVirusCheck = "carbonioAmavisDisableVirusCheck";
+
+    /**
      * Database Custom Urls for freshclam configuration file.
      *
      * @since ZCS 23.1.0
@@ -2721,9 +2729,9 @@ public class ZAttrProvisioning {
     /**
      * Whether the Chats feature enabled for account or COS
      *
-     * @since ZCS 23.1.0
+     * @since ZCS 23.5.0
      */
-    @ZAttr(id=3125)
+    @ZAttr(id=3130)
     public static final String A_carbonioFeatureChatsEnabled = "carbonioFeatureChatsEnabled";
 
     /**
@@ -2749,6 +2757,22 @@ public class ZAttrProvisioning {
      */
     @ZAttr(id=3123)
     public static final String A_carbonioFeatureMailsAppEnabled = "carbonioFeatureMailsAppEnabled";
+
+    /**
+     * Whether the Meeting feature enabled for account or COS
+     *
+     * @since ZCS 23.5.0
+     */
+    @ZAttr(id=3131)
+    public static final String A_carbonioFeatureMeetingEnabled = "carbonioFeatureMeetingEnabled";
+
+    /**
+     * Whether the Team feature enabled for account or COS
+     *
+     * @since ZCS 23.5.0
+     */
+    @ZAttr(id=3132)
+    public static final String A_carbonioFeatureTeamEnabled = "carbonioFeatureTeamEnabled";
 
     /**
      * Logo URL for domain