+++ soap/src/main/java/com/zimbra/soap/account/type/Session.java
@@ -13,16 +13,15 @@ import javax.xml.bind.annotation.XmlValue;
 import com.google.common.base.MoreObjects;
 import com.zimbra.common.soap.HeaderConstants;
 
-import io.leangen.graphql.annotations.GraphQLIgnore;
-import io.leangen.graphql.annotations.GraphQLNonNull;
-import io.leangen.graphql.annotations.GraphQLQuery;
-import io.leangen.graphql.annotations.types.GraphQLType;
+
+
+
+
 
  // See ZimbraSoapContext.encodeSession:
  // eSession.addAttribute(HeaderConstants.A_TYPE, typeStr).addAttribute(HeaderConstants.A_ID, sessionId).setText(sessionId);
 
 @XmlAccessorType(XmlAccessType.NONE)
-@GraphQLType(name="Session", description="A session")
 public class Session {
 
     /**
@@ -30,7 +29,6 @@ public class Session {
      * @zm-api-field-description Session type - currently only set if value is "admin"
      */
     @XmlAttribute(name=HeaderConstants.A_TYPE, required=false)
-    @GraphQLQuery(name="type", description="Session type - currently only set if value is admin")
     private String type;
 
     /**
@@ -38,8 +36,6 @@ public class Session {
      * @zm-api-field-description Session ID
      */
     @XmlAttribute(name=HeaderConstants.A_ID, required=true)
-    @GraphQLNonNull
-    @GraphQLQuery(name="id", description="Session ID")
     private String id;
 
     public Session() {
@@ -48,17 +44,13 @@ public class Session {
     public void setType(String type) { this.type = type; }
     public void setId(String id) { this.id = id; }
     public void setSessionId(String sessionId) { this.id = sessionId; }
-    @GraphQLQuery(name="type", description="Session type - currently only set if value is admin")
     public String getType() { return type; }
-    @GraphQLNonNull
-    @GraphQLQuery(name="id", description="Session ID")
     public String getId() { return id; }
     /**
      * @zm-api-field-tag session-id
      * @zm-api-field-description Session ID  (same as <b>id</b>)
      */
     @XmlValue
-    @GraphQLIgnore
     public String getSessionId() { return id; }
 
     public MoreObjects.ToStringHelper addToStringInfo(MoreObjects.ToStringHelper helper) {