+++ soap/src/main/java/com/zimbra/soap/mail/type/MailCustomMetadata.java
@@ -16,16 +16,11 @@ import javax.xml.bind.annotation.XmlRootElement;
 import com.google.common.base.MoreObjects;
 import com.google.common.collect.Iterables;
 import com.google.common.collect.Lists;
-import com.zimbra.common.gql.GqlConstants;
 import com.zimbra.common.soap.MailConstants;
 import com.zimbra.soap.base.CustomMetadataInterface;
 
-import io.leangen.graphql.annotations.GraphQLQuery;
-import io.leangen.graphql.annotations.types.GraphQLType;
-
 @XmlAccessorType(XmlAccessType.NONE)
 @XmlRootElement(name=MailConstants.E_METADATA)
-@GraphQLType(name=GqlConstants.CLASS_MAIL_CUSTOM_METADATA, description="Custom metadatas")
 public class MailCustomMetadata
 extends MailKeyValuePairs
 implements CustomMetadataInterface {
@@ -48,7 +43,6 @@ implements CustomMetadataInterface {
     @Override
     public void setSection(String section) { this.section = section; }
     @Override
-    @GraphQLQuery(name=GqlConstants.SECTION, description="Section. If absent this indicates that CustomMetadata info is present but there are no sections to report on.")
     public String getSection() { return section; }
 
     public static List <MailCustomMetadata> fromInterfaces(Iterable <CustomMetadataInterface> params) {