+++ store/src/main/java/com/zimbra/cs/mailbox/Mailbox.java
@@ -974,6 +974,25 @@ public class Mailbox implements MailboxStore {
   }
 
   /**
+   * Returns a {@link MailSender} object based on specific domain properties
+   * that can be used to send mails.
+   *
+   * @param domain {@link com.zimbra.cs.account.Domain} to get needed properties
+   * @return {@link MailSender} object
+   * @throws ServiceException if unable to get SMTP session for the current domain
+   *
+   * @author Yuliya Aheeva
+   * @since 23.5.0
+   */
+  public MailSender getMailSender(Domain domain) throws ServiceException {
+    MailSender sender = new MailSender();
+    sender.setTrackBadHosts(true);
+    sender.setSession(domain);
+
+    return sender;
+  }
+
+  /**
    * Returns the list of all <code>Mailbox</code> listeners of a given type. Returns all listeners
    * when the passed-in type is <tt>null</tt>.
    */