+++ store/src/main/java/com/zimbra/cs/util/proxyconfgen/ProxyConfGen.java
@@ -93,7 +93,7 @@ public class ProxyConfGen {
   private static Provisioning mProv = null;
   private static boolean mGenConfPerVhn = false;
   private static boolean hasCustomTemplateLocationArg = false;
-  private static final String CERTBOT_WORKING_DIR = "/common/etc/letsencrypt/live/";
+  private static final String CERTBOT_WORKING_DIR = "/common/certbot/etc/letsencrypt/live/";
   private static final String CERT = "/fullchain.pem";
   private static final String KEY = "/privkey.pem";
 
@@ -2323,8 +2323,11 @@ public class ProxyConfGen {
           new File(mTemplateDir, getConfTemplateFileName("stream.addressBook")),
           new File(mConfIncludesDir, getConfFileName("stream.addressBook")));
       expandTemplate(
-          new File(mTemplateDir, getConfTemplateFileName("stream.chats.messaging.xmpp")),
-          new File(mConfIncludesDir, getConfFileName("stream.chats.messaging.xmpp")));
+          new File(mTemplateDir, getConfTemplateFileName("stream.message.dispatcher.xmpp")),
+          new File(mConfIncludesDir, getConfFileName("stream.message.dispatcher.xmpp")));
+      expandTemplate(
+          new File(mTemplateDir, getConfTemplateFileName("web.clamav.signature.provider")),
+          new File(mConfIncludesDir, getConfFileName("web.clamav.signature.provider")));
       // Templates for ssl mapping
       expandTemplate(
           new File(mTemplateDir, getConfTemplateFileName("map.crt")),
@@ -2483,7 +2486,7 @@ public class ProxyConfGen {
         attrs.put(Provisioning.A_zimbraSSLCertificate, certificate);
         attrs.put(Provisioning.A_zimbraSSLPrivateKey, privateKey);
 
-        LOG.info("Saving " + domainName + " Let's Encrypt certificate/key pair to LDAP.");
+        LOG.info("Saving " + domainName + " Let's Encrypt certificate/key pair to LDAP");
         mProv.modifyAttrs(domain, attrs, true);
 
         entry.sslCertificate = certificate;